fix: rename api -> apis
https://github.com/nextauthjs/next-auth/issues/4986
