+++ app/ranking/page.tsx
@@ -5,7 +5,7 @@ import FullRankingList from '@/components/ranking/fullRankingList';
 import { Box, Flex } from '@radix-ui/themes';
 import styles from './page.module.css';
 import rankingImg from '@/assets/icons/ranking_profile_img.png';
-import { getMemberRanking } from '@/api/rankingApi';
+import { getMemberRanking } from '@/apis/rankingApi';
 import { Student, ApiResponse } from '@/types/rankingType';
 
 export default function Ranking() {
