+++ components/study/SubjectEditForm.tsx
@@ -3,8 +3,8 @@ import React, { useState, useEffect } from 'react';
 import { Text } from '@radix-ui/themes';
 import styles from './SubjectForm.module.css';
 import { useSubjectStore } from '@/store/subjectStore';
-import { postSubjectItem } from '@/api/subjectFormApi';
-import { getSubjectEditList } from '@/api/subjectFormApi';
+import { postSubjectItem } from '@/apis/subjectFormApi';
+import { getSubjectEditList } from '@/apis/subjectFormApi';
 import useSWR from 'swr';
 
 export default function SubjectEditForm({
